move quizList web into modules
