+++ Quiz-web/src/app/app-routing.module.ts
@@ -1,11 +1,8 @@
 import {NgModule} from '@angular/core';
 import {Routes, RouterModule} from '@angular/router';
 
-import {QuizzesComponent} from "./quizzes/quizzes.component";
-
 const routes: Routes = [
   { path:  '', pathMatch:  'full', redirectTo:  'home'},
-  {path: 'api/v1/quizzes', component: QuizzesComponent}
 ];
 
 @NgModule({
