+++ Quiz-web/src/app/app.module.ts
@@ -13,11 +13,11 @@ import {CoreModule} from './modules/core/core.module';
 import {SharedModule} from './modules/shared/shared.module';
 import {AuthRoutingModule} from './modules/authentication/auth-routing.module';
 import {PagesRoutingModule} from './modules/core/pages/pages-routing.module';
-
+import {QuizRoutingModule} from './modules/quiz/quiz-routing.module'
+import {QuizModule} from './modules/quiz/quiz.module'
 @NgModule({
   declarations: [
-    AppComponent,
-    QuizzesComponent
+    AppComponent
   ],
   imports: [
     RouterModule,
@@ -29,10 +29,13 @@ import {PagesRoutingModule} from './modules/core/pages/pages-routing.module';
     CoreModule,
     SharedModule,
     AuthRoutingModule,
-    PagesRoutingModule
+    PagesRoutingModule,
+    QuizRoutingModule,
+    QuizModule
   ],
   providers: [],
   bootstrap: [AppComponent]
 })
+
 export class AppModule {
 }
