+++ Quiz-web/src/app/modules/quiz/quiz-routing.module.ts
@@ -1,0 +1,19 @@
+import {NgModule} from '@angular/core';
+import {Routes, RouterModule} from '@angular/router';
+
+import {QuizzesComponent} from './quizzesList/quizzes.component'
+
+const authenticationRoutes: Routes = [
+  {path: 'quizzes', component: QuizzesComponent}
+];
+
+@NgModule({
+  imports: [
+    RouterModule.forChild(authenticationRoutes)
+  ],
+  exports: [
+    RouterModule
+  ]
+})
+export class QuizRoutingModule {
+}
