+++ Quiz-web/src/app/modules/shared/sidenav/sidenav.component.html
@@ -1,5 +1,5 @@
 
 <div class="sidenav">
-  <a routerLink="api/v1/quizzes">Список вікторин</a>
+  <a routerLink="quizzes">Список вікторин</a>
   <a href="#">Рейтинг учасників</a>
 </div>
